Fixed soundtrack bug and score bug
- Fixed bug where the main theme would play through the game over screen
on the second play through. Theme now just stops playing at the result
screen instead of fading.
- Fixed bug where the score would increment when Ace destroys a robot
after the mirror is destroyed, increasing the score in game but not
increasing the players final score/high score. Now the game does not
increment the score if a robot is destroyed after the mirror has been
destroyed.